Agentic_Apps/pyproject.toml
[project]
name = "simple-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.0",
    "google-genai>=1.47.0",
    "langchain>=1.0.3",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.2",
    "langchain-google-genai>=3.0.0",
    "langchain-groq>=1.0.0",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.2",
    "langgraph-store-mongodb>=0.1.0",
    "ollama>=0.6.0",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
